requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
dash==3.0.4
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2025.5.0
debugpy==1.8.14
decorator==5.2.1
distributed==2025.5.0
distro==1.9.0
emoji==2.14.1
executing==2.2.0
fastjsonschema==2.21.1
fire==0.7.0
Flask==3.0.3
Flask-Login==0.6.3
fsspec==2025.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
locket==1.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
msgpack==1.1.0
narwhals==1.38.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.2.5
openai==1.78.1
packaging==25.0
pandas==2.2.3
parso==0.8.4
partd==1.4.2
platformdirs==4.3.8
plotly==6.0.1
polars==1.29.0
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rpds-py==0.24.0
setuptools==80.3.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
tblib==3.1.0
termcolor==3.1.0
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.0.6
zict==3.0.0
zipp==3.21.0
